End 23/9/2020

+++ app/Http/Controllers/Dashboard/BrandsController.php
@@ -4,18 +4,17 @@ namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\BrandRequest;
-use App\Http\Requests\MainCategoryRequest;
 use App\Models\Brand;
-use App\Models\Category;
-use Illuminate\Http\Request;
 use DB;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class BrandsController extends Controller
 {
 
     public function index()
     {
-         $brands = Brand::orderBy('id','DESC')->paginate(PAGENATION_COUNT);
+        $brands = Brand::orderBy('id', 'DESC')->paginate(PAGENATION_COUNT);
         return view('dashboard.brands.index', compact('brands'));
     }
 
@@ -24,42 +23,36 @@ class BrandsController extends Controller
         return view('dashboard.brands.create');
     }
 
-
     public function store(BrandRequest $request)
     {
 
+        DB::beginTransaction();
 
-            DB::beginTransaction();
+        //validation
 
-            //validation
+        if (!$request->has('is_active')) {
+            $request->request->add(['is_active' => 0]);
+        } else {
+            $request->request->add(['is_active' => 1]);
+        }
 
-            if (!$request->has('is_active'))
-                $request->request->add(['is_active' => 0]);
-            else
-                $request->request->add(['is_active' => 1]);
+        $fileName = "";
+        if ($request->has('photo')) {
 
+            $fileName = uploadImage('brands', $request->photo);
+        }
 
-            $fileName = "";
-            if ($request->has('photo')) {
+        $brand = Brand::create($request->except('_token', 'photo'));
 
-                $fileName = uploadImage('brands', $request->photo);
-            }
+        //save translations
+        $brand->name = $request->name;
+        $brand->photo = $fileName;
 
-            $brand = Brand::create($request->except('_token','photo'));
-
-            //save translations
-            $brand->name = $request->name;
-            $brand -> photo = $fileName;
-
-            $brand->save();
-            DB::commit();
-            return redirect()->route('admin.brands.index')->with(['success' => 'تم ألاضافة بنجاح']);
-
-
-
+        $brand->save();
+        DB::commit();
+        return redirect()->route('admin.brands.index')->with(['success' => 'تم ألاضافة بنجاح']);
 
     }
-
 
     public function edit($id)
     {
@@ -67,13 +60,13 @@ class BrandsController extends Controller
         //get specific categories and its translations
         $brand = Brand::find($id);
 
-        if (!$brand)
-            return redirect()->route('admin.brands.index')->with(['error' => 'هذا القسم غير موجود ']);
+        if (!$brand) {
+            return redirect()->route('admin.brands,index')->with(['error' => 'هذا الماركة غير موجود']);
+        }
 
         return view('dashboard.brands.edit', compact('brand'));
 
     }
-
 
     public function update($id, BrandRequest $request)
     {
@@ -82,31 +75,76 @@ class BrandsController extends Controller
 
             //update DB
 
-
             $brand = Brand::find($id);
 
-            if (!$brand)
-                return redirect()->route('admin.brands.index')->with(['error' => 'هذا القسم غير موجود']);
+            if (!$brand) {
+                return redirect()->route('admin.brands.index')->with(['error' => 'هذا الماركة غير موجود']);
+            }
+
+            DB::beginTransaction();
+
+            /************************************** */
+        //     $data = $request->except('photo');
+
+
+        //     if ($request->hasfile('photo') && $request->file('photo')->isvalid()) {
+        //         $image = $request->file('photo');
+
+        //         if ($brand->photo && Storage::disk('brands')->exists($brand->photo)) { //if image Exist or old image in brands اذا كام المنتج الو صوره وبدك تعدل عليها
+
+        //             $fileName = $image->storeAs('', basename($brand->photo), 'brands'); // Updated new image same name pre-image(disk in brands from file system)
+        //            // dd($fileName);
+        //          }
+        //          else {
+        //            //  dd('ho daaaaaaaa');
+        //             //if not found old image -- save  New image and ne Name To Images Disk
+        //             $fileName = $image->store('', 'brands'); //storge disk
+        //            // dd($fileName);
+        //             }
+
+
+
+        //         $data['photo'] = $fileName;
+
+
+
+        //     }
+        //     $brand->update($data);
+        //    //save translations
+        //     $brand->name = $request->name;
+        //     $brand->save();
+
+            /****************************************** */
+            if ($request->has('photo')) {
+                deleteImage($brand->photo);
+               // Storage::disk('brands')->delete($brand->photo);
+                $fileName = uploadImage('brands', $request->photo);
+                Brand::where('id', $id)
+                    ->update([
+                        'photo' => $fileName,
+                    ]);
+            }
 
             if (!$request->has('is_active'))
                 $request->request->add(['is_active' => 0]);
             else
                 $request->request->add(['is_active' => 1]);
 
-            $brand->update($request->all());
+            $brand->update($request->except('_token', 'id', 'photo'));
 
             //save translations
             $brand->name = $request->name;
             $brand->save();
 
+            DB::commit();
             return redirect()->route('admin.brands.index')->with(['success' => 'تم ألتحديث بنجاح']);
+
         } catch (\Exception $ex) {
 
+            DB::rollback();
             return redirect()->route('admin.brands.index')->with(['error' => 'حدث خطا ما برجاء المحاوله لاحقا']);
         }
-
     }
-
 
     public function destroy($id)
     {
@@ -115,13 +153,22 @@ class BrandsController extends Controller
             //get specific categories and its translations
             $brand = Brand::find($id);
 
-            if (!$brand)
-                return redirect()->route('admin.maincategories.index')->with(['error' => 'هذا القسم غير موجود ']);
+            if (!$brand) {
+                return redirect()->route('admin.brands.index')->with(['error' => 'هذا الماركة غير موجود ']);
+            }
 
-            $brand->delete();
+            if ($brand->photo) {
+                $brand->delete();
+                $brand->translation()->delete();
 
-            return redirect()->route('admin.brands.index')->with(['success' => 'تم  الحذف بنجاح']);
+                // unlink(public_path('images/'.$products->image));//in php native
+                deleteImage($brand->photo);
 
+
+
+               // Storage::disk('brands')->delete($brand->photo); // in laravel to delete image from images disk
+                return redirect()->route('admin.brands.index')->with(['success' => 'تم  الحذف بنجاح']);
+            }
         } catch (\Exception $ex) {
             return redirect()->route('admin.brands.index')->with(['error' => 'حدث خطا ما برجاء المحاوله لاحقا']);
         }
